+++ src/sam/plugins/index.js
@@ -1,3 +1,6 @@
+let { toLogicalID } = require('@architect/utils')
+let { createFunction } = require('@architect/package')
+
 /**
  * Architect Plugins
  * ---
@@ -39,10 +42,46 @@ module.exports = function plugins (inventory, cloudformation, stage, callback) {
 async function exec (inventory, cloudformation, stage) {
   let arc = inventory.inv._project.arc
   let plugins = inventory.inv._project.plugins ? Object.values(inventory.inv._project.plugins) : []
-  return await plugins.reduce(async function reducer (current, plugin) {
+  let pluginNames = inventory.inv._project.plugins ? Object.keys(inventory.inv._project.plugins) : []
+  // compile each plugins CFN modifications
+  let cfn = await plugins.reduce(async function reducer (current, plugin) {
     let run = plugin.package
     let cloudformation = await current
-    if (run) return await run({ arc, cloudformation, stage, inventory })
-    else return Promise.resolve(cloudformation)
+    if (run) {
+      return run({ arc, cloudformation, stage, inventory, createFunction })
+    }
+    else {
+      return Promise.resolve(cloudformation)
+    }
   }, Promise.resolve(cloudformation))
+  // now grab any variable exports from plugins and inject as SSM parameters
+  return pluginNames.reduce(async function reducer (current, pluginName) {
+    let plugin = inventory.inv._project.plugins[pluginName]
+    let varExports = plugin.variables
+    let cloudformation = await current
+    if (varExports) {
+      let vars = await varExports({ arc, cloudformation, stage, inventory })
+      let keys = Object.keys(vars)
+      if (keys && keys.length) {
+        for (let k of keys) {
+          let Value = vars[k]
+          cloudformation.Resources[toLogicalID(`${pluginName}${k}Param`)] = {
+            Type: 'AWS::SSM::Parameter',
+            Properties: {
+              Type: 'String',
+              Name: { 'Fn::Sub': [
+                '/${AWS::StackName}/${pluginName}/${variableName}',
+                {
+                  pluginName,
+                  variableName: k
+                }
+              ] },
+              Value
+            }
+          }
+        }
+      }
+    }
+    return Promise.resolve(cloudformation)
+  }, Promise.resolve(cfn))
 }