+++ test/integration/static/index-test.js
@@ -1,9 +1,9 @@
 let test = require('tape')
 let { join } = require('path')
+let mockFs = require('mock-fs')
 let proxyquire = require('proxyquire')
 let inventory = require('@architect/inventory')
 let { updater } = require('@architect/utils')
-let mockFs = require('mock-fs')
 
 // Necessary to run test solo
 let aws = require('aws-sdk')
@@ -15,7 +15,7 @@ function publish (params, callback) {
   callback(null, params)
 }
 
-let node_modules = { '@architect': { asap: { dist: { 'index.js': 'hi' } } } }
+let node_modules = mockFs.load('node_modules', { recursive: true })
 let staticDeployPath = join(process.cwd(), 'src', 'static', 'index.js')
 let staticDeployMod = proxyquire(staticDeployPath, {
   './publish': publish
@@ -51,7 +51,7 @@ function staticDeploy (t, callback) {
     else {
       params.inventory = result
       staticDeployMod(params, err => {
-        reset() // Must be reset before any tape tests are resolved because mock-fs#201
+        reset()
         callback(err)
       })
     }
@@ -63,7 +63,6 @@ function staticDeploy (t, callback) {
  * - Unfortunately, proxyquire seems to have a nested file folder + `@global` bug, so we can't run this from index
  *   - Instead, we have to run inventory ourselves on each test, which kinda sucks
  * - Also, it'd be nice to test the CloudFormation stackname code path
- *   - However, mock-fs doesn't play nicely with aws-sdk(-mock)
  */
 
 test('Set up env', t => {
@@ -100,7 +99,8 @@ test(`Publish static deploy if @static is defined`, t => {
   let arc = '@app\n an-app\n @static'
   mockFs({
     'app.arc': arc,
-    'public': {}
+    'public': {},
+    node_modules
   })
   staticDeploy(t, err => {
     if (err) t.fail(err)
@@ -134,7 +134,8 @@ test(`Respect prune param`, t => {
   let arc = '@app\n an-app\n @static'
   mockFs({
     'app.arc': arc,
-    'public': {}
+    'public': {},
+    node_modules
   })
   params.prune = true
   staticDeploy(t, err => {
@@ -149,7 +150,8 @@ test(`Respect prune setting in project manifest`, t => {
   let arc = '@app\n an-app\n @static\n prune true'
   mockFs({
     'app.arc': arc,
-    'public': {}
+    'public': {},
+    node_modules
   })
   staticDeploy(t, err => {
     if (err) t.fail(err)
@@ -163,7 +165,8 @@ test(`Respect folder setting in project manifest`, t => {
   let arc = '@app\n an-app\n @static\n folder some-folder'
   mockFs({
     'app.arc': arc,
-    'some-folder': {}
+    'some-folder': {},
+    node_modules
   })
   staticDeploy(t, err => {
     if (err) t.fail(err)
@@ -177,7 +180,8 @@ test(`Respect prefix param`, t => {
   let arc = '@app\n an-app\n @static'
   mockFs({
     'app.arc': arc,
-    'public': {}
+    'public': {},
+    node_modules
   })
   params.prefix = 'some-prefix'
   staticDeploy(t, err => {
@@ -192,7 +196,8 @@ test(`Respect prefix setting in project manifest`, t => {
   let arc = '@app\n an-app\n @static\n prefix some-prefix'
   mockFs({
     'app.arc': arc,
-    'public': {}
+    'public': {},
+    node_modules
   })
   staticDeploy(t, err => {
     if (err) t.fail(err)