+++ test/mocks/app-with-extensions/src/macros/one.js
@@ -1,0 +1,4 @@
+module.exports = function (arc, cfn) {
+  cfn.macroOne = true
+  return cfn
+}